lighing added
Attenuation added, although attenuation values had to be hardocded in shader, passing value not working and is different than used to. Few functions added that are tweaks of rt3d functions for passing in light & mateial structs as well as light position. Needs tidied and figure out why  the passing of the attenuation val isnt working,